% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_event_start}
\alias{create_event_start}
\title{Create a start event column}
\usage{
create_event_start(tibble, event.col)
}
\arguments{
\item{tibble}{A tibble arranged according to the time series.}

\item{event.col}{A string with the name of the column that defines the time series.}
}
\description{
Create a new column which marks the beginning of each time series in a tibble.
The effect is the same as \code{start_event} from \code{itsadug}, but it works
with tibbles.
}
