% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geom_smooth_ci}
\alias{geom_smooth_ci}
\title{Smooths and confidence intervals.}
\usage{
geom_smooth_ci(group = NULL, ci_z = 1.96, ci_alpha = 0.1,
  data = NULL, ...)
}
\arguments{
\item{group}{The optional grouping factor.}

\item{ci_z}{The z-value for calculating the CIs (the default is \code{1.96} for 95 percent CI).}

\item{ci_alpha}{Transparency value of CIs (the default is \code{0.1}).}

\item{data}{The data to be displayed in this layer. If \code{NULL}, it is inherited.}

\item{...}{Arguments passed to \code{geom_path()}.}
}
\description{
It provides a `geom` for plotting GAM smooths with confidence intervals.
}
