% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gam_predictions}
\alias{get_gam_predictions}
\title{Get predictions from a GAM model.}
\usage{
get_gam_predictions(model, time_series, conditions = NULL,
  exclude_random = TRUE)
}
\arguments{
\item{model}{A \code{gam} or \code{bam} model object.}

\item{time_series}{An unquoted expression indicating the model term that defines the time series.}

\item{conditions}{A list of quosures with \link[rlang]{quos} specifying the levels to plot from the model terms.}

\item{exclude_random}{Whether to exclude random smooths (the default is \code{TRUE}).}
}
\description{
It returns a tibble with the predictions from a a \link[mgcv]{gam} or \link[mgcv]{bam} object.
}
