% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inter_df}
\alias{inter_df}
\title{Dataset with two factors}
\format{
A tibble with 1259 observations and 4 variables.
\describe{
\item{\code{x0}}{time series}
\item{\code{y}}{outcome variable}
\item{\code{x1}}{factor with three levels}
\item{\code{x2}}{factor with two levels}
}
}
\usage{
inter_df
}
\description{
A dataset with a normal-distributed outcome variable and two factors.
}
\keyword{datasets}
