% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_difference-defunct}
\alias{plot_difference-defunct}
\title{Plot difference smooth from a GAM.}
\arguments{
\item{difference}{A named list with the levels to compute the difference of.}

\item{conditions}{A named list specifying the levels to plot from the model terms not among \code{series} or \code{difference}. Notice the difference with \link[tidymv]{plot_smooths}, which uses \link[rlang]{quos}.}
}
\description{
It plots the difference smooth from a \link[mgcv]{gam} or \link[mgcv]{bam}.
Significant differences are marked with red areas.
}
\seealso{
\code{\link{tidymv-defunct}}
}
\keyword{internal}
