% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_gami}
\alias{plot_gami}
\title{Plot GAM estimate smooths of an interaction.}
\usage{
plot_gami(model, view, interactions, conditions = NULL, rm_re = FALSE,
  print_summary = TRUE)
}
\arguments{
\item{model}{A \code{gam} or \code{bam} model object.}

\item{view}{A string with the predictor determining the time series.}

\item{interactions}{A named list the levels to be plotted from the main comparison predictor and the interactions.}

\item{conditions}{A named list with rhe values to use for other predictors.}

\item{rm_re}{Whether to remove random effects (the default is \code{FALSE}).}

\item{print_summary}{Wether to print a summary of the values selected for each predictor (the default is \code{TRUE}).}
}
\description{
It plots estimate smooths of the interaction terms from a \link[mgcv]{gam} or \link[mgcv]{bam}.
}
