% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gamsd-defunct}
\alias{plot_gamsd-defunct}
\title{Plot GAM estimate smooths and difference curve.}
\arguments{
\item{model}{A \code{gam} or \code{bam} model object.}

\item{view}{The predictor determining the time series.}

\item{comparison}{The levels for the comparison as a named list.}

\item{conditions}{The values to use for other predictors as a named list.}

\item{rm_re}{Whether to remove random effects (the default is \code{FALSE}).}

\item{bw}{Whether to plot in black and white (the default is \code{FALSE}).}

\item{ylim}{Limits of the y-axis of the smooths panel.}
}
\description{
It plots comparison smooths from the estimates of a \link[mgcv:gam]{gam} or \link[mgcv:bam]{bam}
and the difference curve. Significant differences are marked with red areas.
}
\seealso{
\code{\link[=tidymv-defunct]{tidymv-defunct()}}
}
\keyword{internal}
