% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_gamsd}
\alias{plot_gamsd}
\title{Deprecated function plot_gamsd()}
\usage{
plot_gamsd(...)
}
\arguments{
\item{...}{Not used.}
}
\description{
This function is deprecated and has been removed. Please, use \link[tidymv]{plot_smooths} and  \link[tidymv]{plot_difference}.
}
