% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymv-package.R
\docType{package}
\name{tidymv-package}
\alias{tidymv}
\alias{tidymv-package}
\title{tidymv: Tidy Model Visualisation for Generalised Additive Models}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides functions for visualising generalised
    additive models and getting predicted values using tidy tools from the 'tidyverse' packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stefanocoretta/tidymv}
  \item Report bugs at \url{https://github.com/stefanocoretta/tidymv/issues}
}

}
\author{
\strong{Maintainer}: Stefano Coretta \email{stefano.coretta@gmail.com}

Other contributors:
\itemize{
  \item Jacolien van Rij (Imported functions from itsadug) [copyright holder]
  \item Martijn Wieling (Imported functions from itsadug) [copyright holder]
}

}
\keyword{internal}
