% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_gami}
\alias{plot_gami}
\title{Plot GAM estimate smooths of an interaction.}
\usage{
plot_gami(model, view, interaction, conditions = NULL, rm_re = FALSE)
}
\arguments{
\item{model}{A \code{gam} or \code{bam} model object.}

\item{view}{The predictor determining the time series.}

\item{interaction}{The levels of the interaction terms as a named list.}

\item{conditions}{The values to use for other predictors as a named list.}

\item{rm_re}{Whether to remove random effects (the default is \code{FALSE}).}
}
\description{
It plots estimate smooths of the interaction terms from a \link[mgcv]{gam} or \link[mgcv]{bam}.
}
