% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_smooths}
\alias{plot_smooths}
\title{Plot GAM smooths.}
\usage{
plot_smooths(model, time_series, comparison, facet_terms = NULL,
  conditions = NULL, exclude_random = TRUE, exclude_terms = NULL,
  series_length = 25, split = NULL, sep = "[^[:alnum:]]+")
}
\arguments{
\item{model}{A \code{gam} or \code{bam} model object.}

\item{time_series}{An unquoted expression indicating the model term that defines the time series.}

\item{comparison}{An unquoted expression indicating the model term for which the comparison will be plotted.}

\item{facet_terms}{An unquoted formula with the terms used for faceting.}

\item{conditions}{A list of quosures with \link[rlang]{quos} specifying the levels to plot from the model terms not among \code{time_series}, \code{comparison}, or \code{facet_terms}.}

\item{exclude_random}{Whether to exclude random smooths (the default is \code{TRUE}).}

\item{exclude_terms}{Terms to be excluded from the prediction. Term names should be given as they appear in the model summary (for example, \code{"s(x0,x1)"}).}

\item{series_length}{An integer indicating how many values along the time series to use for predicting the outcome term.}
}
\description{
It plots the smooths from the estimates of a \link[mgcv]{gam} or \link[mgcv]{bam} object.
}
