% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_smooths}
\alias{plot_smooths}
\title{Plot GAM smooths.}
\usage{
plot_smooths(model, time_series, comparison, facet_terms, conditions = NULL,
  plot_random = FALSE)
}
\arguments{
\item{model}{A \code{gam} or \code{bam} model object.}

\item{time_series}{An unquoted expression indicating the model term that defines the time series.}

\item{comparison}{An unquoted expression indicating the model term for which the comparison will be plotted.}

\item{facet_terms}{An unquoted formula with the terms used for faceting.}

\item{conditions}{A list of quosures with \link[rlang]{quos} with the conditions for plotting if \code{facet_terms} is \code{NULL}.}

\item{plot_random}{Whether to plot random smooths (the default is \code{FALSE}).}
}
\description{
It plots the smooths from the estimates of a \link[mgcv]{gam} or \link[mgcv]{bam} object.
}
