% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R, R/tidymv-defunct.R
\name{plot_gamsd}
\alias{plot_gamsd}
\alias{tidymv-defunct}
\title{Deprecated functions in package \pkg{tidymv}.}
\usage{
plot_gamsd(...)
}
\description{
The functions listed below are defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for defunct functions are
  available at \code{help("<function>-defunct")}.
}
\section{This function is deprecated and has been removed. Please, use \link[tidymv]{plot_smooths} and  \link[tidymv]{plot_difference}.}{
NA
}

\keyword{internal}
