% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymv.R
\docType{package}
\name{tidymv}
\alias{tidymv}
\title{tidymv: Plotting for generalised additive models.}
\description{
This package provides functions for visualising generalised additive models
and get predicted values using tidy tools from the tidyverse. The name stands
for TIDY Model Visualisation.
}
